+++ projects/egeo/src/lib/st-form/st-form.model.ts
@@ -11,13 +11,15 @@
 
 
 import { JSONSchema4 } from 'json-schema';
-import { FORM_UI_COMPONENT } from './shared/ui-component.interface';
+import { StDropDownMenuItem } from '..';
+import { FORM_UI_COMPONENT } from './st-form-field/st-form-field.interface';
 
 export interface StFormUIDefinition {
    relatedTo?: string;
    visible?: { [key: string]: any };
    component?: FORM_UI_COMPONENT;
    link?: string;
+   options?: StDropDownMenuItem[];
 }
 
 export interface StFormSchema extends JSONSchema4 {