+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 * st-table: Filter menus are displayed aligned to the left side of their cell
 * st-checkbox: Allow to customize checked icon from outside using css variables
 * st-zero-page: Allow to customize some styles from outside using css variables
+* st-form: Allow to render selects with integer value
+* st-form: Allow to customize select option labels using the UI Definition object
 
 **Fixed bugs:**